Renamed type module to ty as per convention e.g. syn
